+++ snake_ml/system/drivers/OSDriver.h
@@ -28,11 +28,11 @@ public:
 	virtual void LogMessage(std::wstring msg) const = 0;
 
 	Event<IOSDriver, uint64_t>	m_onInputEvent;
-	Event<IOSDriver, double>	m_onUpdateEvent;
+	Event<IOSDriver, float>	m_onUpdateEvent;
 
 protected:
 	void SendInputEvent(uint64_t);
-	void SendUpdateEvent(double);
+	void SendUpdateEvent(float);
 
 	virtual void OnInitialize() = 0;
 	virtual void OnRunOSMainLoop() = 0;
