+++ snake_ml/system/ecs/SystemsPool.h
@@ -15,7 +15,7 @@ public:
 	void ScheduleSystem(std::unique_ptr<ISystem>&& system);
 	void UnscheduleSystem(const ISystem* system);
 
-	void Update(double deltaTime);
+	void Update(float deltaTime);
 
 private:
 	std::vector<std::unique_ptr<ISystem>> m_systems;
